Refacter code so flashcard for mis reused for editing a flashcard #15
